Created "nested" partial object when invoked with '...'
When calling a partial object (to supply the remaining arguments and
complete the function application - i.e. perform the actual function
call), interpret an Ellipsis in the argument list as another (nested)
partial application.

In other words, given

  def foo(a, b, c):
      return a + b + c

  bar = partial(foo, 1)

this change allows

  baz = bar(2, ...)

to be equivalent to

  baz = partial(bar, 2)

both of which allows

  baz(3)  # returns 6

The same applies to partially applying arguments from the right side:

  bar = partial(foo, ..., 3)
  baz = bar(..., 2)
  baz(1)  # return 6